// src/app/models/billingModel.ts
export interface BillingModel {
    company: string;
    country: string;
    address: string;
    address2: string;
    city: string;
    province: string;
    zip: number;
    phone: string;
}

export class BillingModel {
    company: string;
    country: string;
    address: string;
    address2: string;
    city: string;
    province: string;
    zip: number;
    phone: string;
}